+++ HISTORY.rst
@@ -4,6 +4,7 @@ History
 
 Pre-release
 -----------
+* Soils now define water mass density `wmd` and `liq_sg` instead of `liq_mass_density` to allow for not water liquids.
 * Can set min shear wave velocity in sra using `vs_min`.
 
 0.6.18 (2020-12-08)